streamlit_interface/requirements.txt
pillow==9.2.0
streamlit==1.12.0
python-dotenv==0.20.0
